✨ feat: Add babel-present-everything.png and update README.md
The commit adds the babel-present-everything.png file and updates the README.md file to include a reference to the newly added image.

Note: The conventional commits specification recommends using imperative mood in the subject line, but in this case, since it's referring to an action that has already been done (adding the file and updating the README), the present tense is used.